Move bookParser test scenarios to file
Makes it easier to add future test cases
